admin part done

+++ app/Http/Controllers/ProductController.php
@@ -10,6 +10,17 @@ use Illuminate\Support\Facades\Hash;
 
 class ProductController extends Controller
 {
+    private $pictureFolder;
+
+    /**
+     * ProductController constructor.
+     * @param $pictureFolder
+     */
+    public function __construct()
+    {
+        $this->pictureFolder = public_path('storage/img/products/');
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -42,12 +53,39 @@ class ProductController extends Controller
     {
         $request->validate([
             'name' => 'required|min:5|max:100',
-            'description' => 'required',
+            'description' => 'required|min:5',
+            'reference' => 'required|min:5|max:16',
+            'category' => 'required',
             'size' => 'required',
+            'code' => 'required',
+            'status' => 'required',
             'product_picture' => 'image',
-            'price' => 'numeric|deci'
+            'price' => 'numeric|required'
         ]);
-        //
+
+        $product = Product::create($request->all());
+        $category = Category::find($request->input('category'));
+        $product->category()->associate($category);
+
+        if ($request->file('product_picture') != null) {
+            $newFile = $request->file('product_picture')->getPathname();
+            $originalName = $request->file('product_picture')->getClientOriginalName();
+            $fileExtension = explode(".", $originalName)[1];
+            $newFileName = substr(Hash::make($product->name), 0, 20);
+            $pictureLink = $newFileName . "." . $fileExtension;
+
+            $picture = new Picture();
+            trim(str_replace([" ", ".", "/", "'", "~"], "", $newFileName));
+            $picture->link = $pictureLink;
+            $picture->name = $request->input('name') . "_image";
+            copy($newFile, $this->pictureFolder . $category->name . '/' . $pictureLink);
+
+            $picture->category()->associate($category);
+            $picture->product()->associate($product);
+            $picture->save();
+        }
+        $product->save();
+        return redirect()->route('products.index')->with('message', 'Le produit ' . strtoupper($product->name) . ' a bien été créé');
     }
 
     /**
@@ -75,11 +113,10 @@ class ProductController extends Controller
     }
 
     /**
-     * Update the specified resource in storage.
-     *
-     * @param \Illuminate\Http\Request $request
-     * @param int $id
-     * @return \Illuminate\Http\Response
+     * @param Request $request
+     * @param $id
+     * @return \Illuminate\Http\RedirectResponse
+     * @throws \Exception
      */
     public function update(Request $request, $id)
     {
@@ -94,6 +131,7 @@ class ProductController extends Controller
         ]);
 
         $product = Product::find($id);
+        $product->update($request->all());
         $category = Category::find($request->input('category'));
         //updating category
         if ($product->category_id != $request->input('category')) {
@@ -103,44 +141,41 @@ class ProductController extends Controller
 
         //updating picture
         $picture = $product->picture;
-        $productsPicturesFolder = public_path('storage/img/products/');
         if ($request->file('product_picture') != null) {
+
             $newFile = $request->file('product_picture')->getPathname();
             $originalName = $request->file('product_picture')->getClientOriginalName();
             $fileExtension = explode(".", $originalName)[1];
-            $newFileName = substr(Hash::make($product->name), 0, 20);
+            $newFileName = $product->id . $product->reference;
+            $pictureLink = $newFileName . "." . $fileExtension;
 
-            if ($picture != null) {
-                if ($picture->category_id != $request->input('category')) {
-                    $picture->category()->dissociate();
-                }
-            } else {
+            if ($picture == null) {
                 $picture = new Picture();
+            } else {
+                if (file_exists($this->pictureFolder . $category->name . '/' . $picture->link) && !is_dir($this->pictureFolder . $category->name . '/' . $picture->link)) {
+                    unlink($this->pictureFolder . $category->name . '/' . $picture->link);
+                }
             }
 
-            if (file_exists($productsPicturesFolder . $category->name . '/' . $picture->link) && !is_dir($productsPicturesFolder . $category->name . '/' . $picture->link)) {
-                unlink($productsPicturesFolder . $category->name . '/' . $picture->link);
-            }
             trim(str_replace([" ", ".", "/", "'", "~"], "", $newFileName));
-            $picture->link = $newFileName . "." . $fileExtension;
-            $picture->name = $product->name . "_image";
-            copy($newFile, $productsPicturesFolder . $category->name . '/' . $picture->link);
+            $picture->link = $pictureLink;
+            $picture->name = $request->input('name') . "_image";
+            copy($newFile, $this->pictureFolder . $category->name . '/' . $pictureLink);
 
             $picture->category()->associate($category);
             $picture->product()->associate($product);
             $picture->save();
-        } else {
-            if ($picture != null) {
-                $picture->product()->dissociate();
-                $picture->category()->dissociate();
-                $picture->delete();
-                unlink($productsPicturesFolder . $category->name . '/' . $picture->link);
-            }
         }
 
-        $product->update($request->all());
-        // $product->save();
-        return redirect()->route('products.index')->with('message', 'Le produit ' . $product->name . ' a bien été mis à jour');
+        if ($picture != null && $request->input('picture_src') == null) {
+            $picture->product()->dissociate();
+            $picture->category()->dissociate();
+            $picture->delete();
+            unlink($this->pictureFolder . $category->name . '/' . $picture->link);
+        }
+
+        $product->save();
+        return redirect()->route('products.index')->with('message', 'Le produit ' . strtoupper($product->name) . ' a bien été mis à jour');
     }
 
     /**
@@ -151,6 +186,21 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
-        //
+        $product = Product::find($id);
+        $name = strtoupper($product->name);
+        $picture = $product->picture;
+
+        if ($picture != null) {
+            $image = $this->pictureFolder . $product->category->name . '/' . $picture->link;
+            if (file_exists($image) && !is_dir($image)) {
+                $picture->product()->dissociate();
+                $picture->delete();
+                unlink($image);
+            }
+        }
+        $product->category()->dissociate();
+        $product->delete();
+
+        return redirect()->route('products.index')->with('message', 'Le produit ' . strtoupper($name) . ' a été supprimé');
     }
 }
